renderer: serialize draw operations
We previously allowed the CPU to modify the shared resources async with the
GPU because of a laissez-faire attitude towards dropped/torn frames.

However, it is not just misrendered frames that is the issue. Because our
CSRenderSource properties are non-atomic, it could be the case that the CPU
re-creates a texture/buffer, releases the old object, and at the same time
the renderer thread gets a references to the now stale object.

To solve these memory issues, we ensure that the CSRenderer is serialized
with all draw operations. We also set the MTKView to explicit draw mode
which should only render when needed.