display: use blit functions for canvas mode
Previously, in non-GL mode, we would copy the canvas pixel data with the CPU
into the texture to render. This was not synced with the GPU and so it was
possible for undefined accesses to happen.

Now, we will define a MTLBuffer over the canvas pixel data. Then we call the
GPU blit function to copy that data into a private texture. While this
prevents undefined accesses, tearing is still possible if the underlying
canvas data changes from the point where we invalidate a rectangle to the
point where that data is copied to the GPU. However, SPICE does not provide
any way to prevent this (as opposed to in GL-mode) and in practice we will
be able to render a lot faster than the VM guest can produce display data.