display: re-work GL draw flow
We require every GL draw request from SPICE to respond with GL done after
the render completes in order to synchronize drawing with the guest.

We also cannot send extra GL done when, for example, a render pass is done
because of a cursor movement. The old way to handle this was akward: we
store a variable indicating we need to send a GL done and send it when we
get a callback from the renderer.

The new way is to pass the callback directly. This also improves the code
quality.