renderer: draw to a backbuffer first
Another attempt at re-arch of the renderer. This time, we allow any number
of draws directly to a backbuffer. The main advantage is decoupling of the
view drawing which had a lot of complicated inter-locking mechnisms and
resulted in issues such as slow rendering of GRUB due to iOS limiting the
drawable allocation to 60 FPS. The main disadvantage is that we now add
extra memory usage for the back-buffer, we waste a lot of GPU cycles drawing
updates that may never be seen, and introduce some latency due to the added
render pass. However, all of that may not amount to much compared to the
~100ms delay imposed by the 60 FPS limit.