display: use private queue for draw operations
Since moving to rendering to a backbuffer, we no longer need to keep
CSDisplay and CSRenderer in sync (and CSCursor as well). This simplifies
much of the logic in concurrency. We have a couple of points where two
queues must be in sync: `cs_cursor_invalidate` for CSCursor <-> CSDisplay
and CSRenderSourceDelegate interface for CSDisplay <-> CSRenderer.