renderer: remove need for back buffer
With all the changes made to support rendering to a back buffer, we realized
that the back buffer is actually not needed now that proper synchronization
is taking place. Instead of drawing to a backbuffer, all we have to do is
keep a copy of the CSRenderSource (this is important). When using the
canvas, we can batch together all the blit requests (like what happened
before the last change) and have it scheduled to run when the MTKView draw
event comes in. Third time's the charm!