renderer: move blit back to CPU by default
There are some performance issues that lead to tearing when canvas rendering
is used on non-Apple Silicon devices and many blit operations happen in a
short period of time. This is likely due to bandwidth issues. On Apple
Silicon, with unified memory, there is no advantage of GPU blit anyways.
We default back to CPU bliting unless the client specifically requests GPU.