display: mitigate deadlock on scanout texture
A race on the server side could allow the same pipe to be sent multiple
times causing the read() to block. This mitigates that situations but does
not completely fix the fact that an older scanout texture is actually the
most recent. However, by the time we discover this, we might have already
thrown that out for a stale texture that was sent after. This hopefully will
not be an issue because the IOSurfaceLookup would fail both times but if the
resolution happened to change in between, we could get bad data until the
next scanout. All of this would be properly mitigated when we implement a
better IPC solution than the current hack which is sending the global
IOSurface ID through a pipe.

Fixes #5