main: fix use after free from forgetting to capture self
We use CSMain.syncWith in dealloc to release the GObjects but
when we use CSMain.asyncWith elsewhere, we need to make sure
to capture self so it doesn't get deallocated while the async
call is outstanding.

Also fix another subtle bug when setting spiceMain property
because we might have a stale variable.